requirements.txt
opencv-contrib-python==4.8.1.78
numpy==1.24.2
Pillow==10.4.0
imutils==0.5.4
PyQt5==5.15.10
scipy==1.10.1
pyaudio==0.2.13